Add new credentials and a script to sign commmits.
We need the key, the passphrase and the id for the script which signs
commits.

There is a git-sign script so we can call `git sign` which imports the
key from the environment variable which will be passed in from the job.

The key is then removed after use so it isn't stored on Jenkins. This is
less of a problem with the node images as they are destroyed after use
but it's worth doing.

The passphrase is passed as a PASSPHRASE environment variable which is
read by gpg when it's signing the commit.

The Jenkinsfile will look a bit like this:

```
withCredentials([
 string(credentialsId: 'github-gpg-key', variable: 'GPG_KEY'),
 string(credentialsId: 'github-gpg-passphrase', variable: 'PASSPHRASE'),
 string(credentialsId: 'github-gpg-key-id', variable: 'GPG_KEY_ID')
 ]) {
 ..git commands
```